Feat | Hook useWeather & new api
